refactor(db): rename neo4j-sync → graph-sync (writes Postgres, not Neo4j) (BI-2A3BE4D7)
BET-5 Phase-1: since BET-5, packages/db/src/neo4j-sync.ts writes the Postgres
graph mirror (graph_node/graph_edge via prisma.$executeRawUnsafe) with zero Cypher
— the "neo4j" name is now actively misleading (it caused a false "EA writes broken"
alarm on 2026-07-15). Rename the file and its public @dpf/db subpath to graph-sync.

Scope (neo4j-sync only; neo4j-schema deferred to Phase 3 — it still holds the
Neo4j-specific initNeo4jSchema/init-neo4j dev seed removed with the legacy drivers):
- git mv neo4j-sync.ts→graph-sync.ts, neo4j-sync.test.ts→graph-sync.test.ts,
  test/neo4j-sync-scope.test.ts→graph-sync-scope.test.ts.
- Public API: packages/db/package.json "./neo4j-sync"→"./graph-sync"; every consumer
  updated INCLUDING the dynamic string refs tsc can't see —
  apps/web/lib/actions/ea.test.ts vi.mock + await import("@dpf/db/graph-sync").
- index.ts re-export; internal imports (neo4j-rebuild-*, discovery-*); apps/web
  ea.ts + documents/document-store.ts.
- EA SysML seed sourceKeys in seed-ea-sysml-data-authority.ts repointed to
  graph-sync.ts so the authority-contract records don't go stale.

No behavior change. Validated: 13/13 @dpf/db graph-sync + scope tests; 25/25
apps/web ea.test.ts (proves the dynamic subpath resolves); packages/db tsc 0 errors;
apps/web tsc clean. No "neo4j-sync" references remain.

Local-CI-Override: db+web tsc clean at 8GB + 38 unit tests green; local prepush tsc OOMs cold. CI authoritative.
Process-Spine-Decision: extend — BET-5 Phase-1 naming cleanup; no behavior/spine change.
Design-Grounding-Decision: extend — apps/web import-path + EA seed update only; no UI/contract change.
BI: BI-2A3BE4D7
