+++ packages/db/src/seed-ea-sysml-data-authority.ts
@@ -38,9 +38,9 @@ const arch = (extra: Record<string, unknown> = {}) => ({ provenance: "architect"
 
 // ── Parts (entities) → realizing substrate ──────────────────────────────────
 const PARTS: Array<ElementSeed & { satisfies: string[] }> = [
-  { key: "part:postgres",   typeSlug: "part_definition", name: "Postgres (System of Record)", description: "The single authority for all entities; every application/seed write lands here first. Authority is a design contract (one-way projection writers), not a DB-level guard.", properties: arch({ sourceKey: "packages/db/src/neo4j-sync.ts#L2-4 (\"Postgres is always the authority\")" }), satisfies: [] },
+  { key: "part:postgres",   typeSlug: "part_definition", name: "Postgres (System of Record)", description: "The single authority for all entities; every application/seed write lands here first. Authority is a design contract (one-way projection writers), not a DB-level guard.", properties: arch({ sourceKey: "packages/db/src/graph-sync.ts#L2-4 (\"Postgres is always the authority\")" }), satisfies: [] },
   { key: "part:prisma",     typeSlug: "part_definition", name: "Prisma Schema (canonical definitions)", description: "schema.prisma — the source-of-truth structural definition of the authority; read by the mirror via schema-source.ts.", properties: det("packages/db/prisma/schema.prisma; packages/db/src/schema-source.ts"), satisfies: [] },
-  { key: "part:neo4j",      typeSlug: "part_definition", name: "Neo4j Projection (graph)", description: "Derived graph projection of the EA graph, infra/IT4IT topology, and document references. Fire-and-forget MERGE writes that never throw — a projection outage cannot block a Postgres write.", properties: det("packages/db/src/neo4j-sync.ts"), satisfies: ["req:REQ-DATA-1"] },
+  { key: "part:neo4j",      typeSlug: "part_definition", name: "Neo4j Projection (graph)", description: "Derived graph projection of the EA graph, infra/IT4IT topology, and document references. Fire-and-forget MERGE writes that never throw — a projection outage cannot block a Postgres write.", properties: det("packages/db/src/graph-sync.ts"), satisfies: ["req:REQ-DATA-1"] },
   { key: "part:qdrant",     typeSlug: "part_definition", name: "Qdrant Projection (vector memory)", description: "Derived vector projection for semantic recall: four 768-dim Cosine collections (agent-memory, platform-knowledge, wiki-pages, documents). Recall reads it for relevance; Postgres stays authoritative (e.g. commandments return even when Qdrant is down).", properties: det("packages/db/src/qdrant.ts; apps/web/lib/inference/embedding.ts"), satisfies: ["req:REQ-DATA-1"] },
   { key: "part:mirror",     typeSlug: "part_definition", name: "Data-Model Mirror / Reconciler", description: "Deterministic, idempotent Prisma→EA projection: parse schema → pure diff (planMirror) → apply (reconcileDataModelMirror) → best-effort Neo4j sync. EP-DATA-ARCH.", properties: det("apps/web/lib/ea/data-model-mirror.ts; apps/web/lib/ea/data-model-mirror-apply.ts; apps/web/lib/ea/run-data-model-mirror.ts"), satisfies: ["req:REQ-DATA-2", "req:REQ-DATA-4", "req:REQ-DATA-5"] },
   { key: "part:dataobject", typeSlug: "part_definition", name: "DataObject Mirror Nodes", description: "The mirror's output in the EA substrate: one data_object EaElement per Prisma model + one associated_with EaRelationship per relation, each carrying a stable source key.", properties: det("packages/db/prisma/schema.prisma#EaElement; apps/web/lib/ea/data-model-mirror.ts#sourceKey"), satisfies: ["req:REQ-DATA-3"] },
@@ -51,7 +51,7 @@ const PARTS: Array<ElementSeed & { satisfies: string[] }> = [
 
 // ── Requirements (invariants) ───────────────────────────────────────────────
 const REQUIREMENTS: ElementSeed[] = [
-  { key: "req:REQ-DATA-1", typeSlug: "requirement", name: "REQ-DATA-1 Projections are derived, never authoritative", description: "No projection (Neo4j, Qdrant, EA mirror) writes back to Postgres. Projection writers are one-way and fail-soft (log, never throw); recall always treats Postgres as authority.", properties: arch({ note: "design contract (one-way writers), not a DB constraint", sourceKey: "packages/db/src/neo4j-sync.ts#L2-4; apps/web/lib/inference/embedding.ts" }) },
+  { key: "req:REQ-DATA-1", typeSlug: "requirement", name: "REQ-DATA-1 Projections are derived, never authoritative", description: "No projection (Neo4j, Qdrant, EA mirror) writes back to Postgres. Projection writers are one-way and fail-soft (log, never throw); recall always treats Postgres as authority.", properties: arch({ note: "design contract (one-way writers), not a DB constraint", sourceKey: "packages/db/src/graph-sync.ts#L2-4; apps/web/lib/inference/embedding.ts" }) },
   { key: "req:REQ-DATA-2", typeSlug: "requirement", name: "REQ-DATA-2 Drift surfaces as a conformance issue, never a silent overwrite", description: "On a duplicate source key the mirror plan is blocked and writes a mirror-duplicate-source-key issue (severity error) WITHOUT writing; steward drift findings are written as issues, not applied as mutations.", properties: det("apps/web/lib/ea/data-model-mirror-apply.ts#writeDuplicateIssues; apps/web/lib/ea/data-architecture-steward.ts") },
   { key: "req:REQ-DATA-3", typeSlug: "requirement", name: "REQ-DATA-3 Every mirrored node carries a stable source key", description: "Element key prisma:model:<Model>; relationship key prisma:relation:<From>:<field>:<To>. Stored in properties.sourceKey and the infraCiKey column, denormalized onto the Neo4j node.", properties: det("apps/web/lib/ea/data-model-mirror.ts#L23-37; apps/web/lib/ea/data-model-mirror-apply.ts#L210") },
   { key: "req:REQ-DATA-4", typeSlug: "requirement", name: "REQ-DATA-4 Mirror is deterministic and idempotent", description: "A no-delta run performs no writes and no snapshot (status noop); the nightly task runs the reconcile directly without an LLM loop.", properties: det("apps/web/lib/ea/data-model-mirror-apply.ts; apps/web/lib/actions/agent-task-scheduler.ts#L200-205") },
@@ -61,7 +61,7 @@ const REQUIREMENTS: ElementSeed[] = [
 
 // ── Constraint ───────────────────────────────────────────────────────────────
 const CONSTRAINTS: ElementSeed[] = [
-  { key: "con:CON-DATA-1", typeSlug: "constraint", name: "CON-DATA-1 One-way authority→projection", description: "dataflow(Postgres) → {Neo4j, Qdrant, EA mirror} only; ∄ edge projection→Postgres. Projection failure ⇒ degrade, never block the authority write.", properties: det("packages/db/src/neo4j-sync.ts", { kind: "parametric" }) },
+  { key: "con:CON-DATA-1", typeSlug: "constraint", name: "CON-DATA-1 One-way authority→projection", description: "dataflow(Postgres) → {Neo4j, Qdrant, EA mirror} only; ∄ edge projection→Postgres. Projection failure ⇒ degrade, never block the authority write.", properties: det("packages/db/src/graph-sync.ts", { kind: "parametric" }) },
 ];
 
 // ── Verification cases (evidence records) ────────────────────────────────────
